clean up parquet schema for storing variants #719
Task-Url: https://github.com/seqpipe/gpf/issues/719

+++ DAE/variants/vcf_utils.py
@@ -9,7 +9,7 @@ import numpy as np
 from itertools import izip
 
 
-GENOTYPE_TYPE = np.int64
+GENOTYPE_TYPE = np.int8
 
 
 def mat2str(mat, col_sep="", row_sep="/"):
